app/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
